Resistance value
A resistor is a circuit device designed to have a specific resistance value between its ends. Resistance values are expressed in ohms or kilo-ohms. Resistors are frequently marked with colored bands that encode their resistance values, as shown in figure below. The first two bands are digits, and the third is a power-of-ten multiplier. Write a program that prompts for the colors of Band 1, Band 2, and Band 3, and then displays the resistance in kilo-ohms. Include a helper function search that takes three parameters—the list of strings, the size of the list, and a target string, and returns the subscript of the list element that matches the target or returns –1 if the target is not in the list.